feedback: move tryStaticEvaluation to js-ast-utils
