+++ packages/@romejs/compiler/transforms/compile/transpile/enums.ts
@@ -1,18 +1,17 @@
 import {
 	AnyJSExpression,
 	AnyJSStatement,
-	JSBinaryExpression,
 	JSBindingIdentifier,
 	JSCallExpression,
 	JSFunctionExpression,
-	JSUnaryExpression,
 	TSEnumDeclaration,
 	jsNumericLiteral,
 	jsStringLiteral,
 } from "@romejs/ast";
 import {LetBinding, Path, VarBinding} from "@romejs/compiler";
 import {REDUCE_REMOVE} from "@romejs/compiler/constants";
-import {template} from "@romejs/js-ast-utils";
+import {descriptions} from "@romejs/diagnostics";
+import {template, tryStaticEvaluation} from "@romejs/js-ast-utils";
 
 interface PreviousEnumMembers {
 	[name: string]: number | string;
@@ -36,11 +35,9 @@ function buildEnumMember(
 	name: string,
 	value: AnyJSExpression,
 ): AnyJSExpression {
-	return (isString ? buildStringAssignment : buildNumericAssignment)(
-		id,
-		name,
-		value,
-	);
+	return isString
+		? buildStringAssignment(id, name, value)
+		: buildNumericAssignment(id, name, value);
 }
 
 function buildNumericAssignment(
@@ -90,7 +87,10 @@ function translateEnumValues(
 			member.id.type === "JSIdentifier" ? member.id.name : member.id.value;
 
 		if (initializer) {
-			const constValue = evaluate(initializer, seen);
+			let {value: constValue, bailed} = tryStaticEvaluation(initializer);
+			if (bailed && initializer.type === "JSReferenceIdentifier") {
+				constValue = seen[initializer.name];
+			}
 			if (constValue !== undefined) {
 				seen[name] = constValue;
 				if (typeof constValue === "number") {
@@ -118,105 +118,21 @@ function translateEnumValues(
 	});
 }
 
-function evaluate(
-	expr: AnyJSExpression,
-	seen: PreviousEnumMembers,
-): number | string | undefined {
-	return evalConstant(expr);
-
-	function evalConstant(expr: AnyJSExpression): number | string | undefined {
-		switch (expr.type) {
-			case "JSStringLiteral":
-				return expr.value;
-			case "JSUnaryExpression":
-				return evalUnaryExpression(expr);
-			case "JSBinaryExpression":
-				return evalBinaryExpression(expr);
-			case "JSNumericLiteral":
-				return expr.value;
-			case "JSReferenceIdentifier":
-				return seen[expr.name];
-			case "JSTemplateLiteral": {
-				if (expr.quasis.length === 1) {
-					return expr.quasis[0].cooked;
-				}
-				return undefined;
-			}
-			default:
-				return undefined;
-		}
-	}
-
-	function evalUnaryExpression(
-		expr: JSUnaryExpression,
-	): number | string | undefined {
-		const value = evalConstant(expr.argument);
-		if (value === undefined) {
-			return;
-		}
-
-		switch (expr.operator) {
-			case "+":
-				return value;
-			case "-":
-				return -value;
-			case "~":
-				return ~value;
-			default:
-				return undefined;
-		}
-	}
-
-	function evalBinaryExpression(expr: JSBinaryExpression): number | undefined {
-		const left = Number(evalConstant(expr.left));
-		if (left === undefined) {
-			return;
-		}
-		const right = Number(evalConstant(expr.right));
-		if (right === undefined) {
-			return;
-		}
-
-		switch (expr.operator) {
-			case "|":
-				return left | right;
-			case "&":
-				return left & right;
-			case ">>":
-				return left >> right;
-			case ">>>":
-				return left >>> right;
-			case "<<":
-				return left << right;
-			case "^":
-				return left ^ right;
-			case "*":
-				return left * right;
-			case "/":
-				return left / right;
-			case "+":
-				return left + right;
-			case "-":
-				return left - right;
-			case "%":
-				return left % right;
-			default:
-				return undefined;
-		}
-	}
-}
-
 export default {
 	name: "enums",
 	enter(path: Path) {
-		const {node} = path;
+		const {context, node} = path;
 
 		if (node.type !== "TSEnumDeclaration") {
 			return node;
 		}
 
 		if (node.const) {
-			throw new Error('"const" enums are not supported');
+			context.addNodeDiagnostic(
+				node,
+				descriptions.COMPILER.CONST_ENUMS_UNSUPPORTED,
+			);
+			return REDUCE_REMOVE;
 		}
 
 		if (node.declare) {
