+++ packages/@romejs/js-ast-utils/index.ts
@@ -41,3 +41,4 @@ export {default as createMemberProperty} from "./createMemberProperty";
 export {default as hasJSXAttribute} from "./hasJSXAttribute";
 export {default as getJSXAttribute} from "./getJSXAttribute";
 export {default as isJSXElement} from "./isJSXElement";
+export {default as tryStaticEvaluation} from "./tryStaticEvaluation";
